% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{\%as\%}
\alias{\%as\%}
\alias{as}
\title{Abbreviate a package}
\usage{
pkg \%as\% nm
}
\arguments{
\item{pkg}{A name of package}

\item{nm}{An abbreviated name of the package}
}
\description{
Abbreviate a package
}
\examples{
  graphics \%as\% gr
  gr$plot

}
