% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_as.R
\name{import_as}
\alias{import_as}
\title{Load packages into an aliased namespace.}
\usage{
import_as(package, alias)
}
\arguments{
\item{package}{The name of the package to be loaded.}

\item{alias}{The name of the namespace to load the package into.}
}
\value{
Nothing.
}
\description{
Register a new namespace with alias. Then, loads a package into
that namespace. Allows namespace management similar to python.
}
