% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-importas.R
\name{package}
\alias{package}
\title{Import a package}
\usage{
package(pkg)
}
\arguments{
\item{pkg}{A package's name}
}
\value{
An \code{importas}-class object. It is a list of \code{NULL}s whose names
correspond to those of exported objects from a specified package.
}
\description{
Import a package
}
