% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{::}
\alias{::}
\title{Overwrite :: Operator}
\usage{
pkg::name
}
\arguments{
\item{pkg}{package name: symbol or literal character string.}

\item{name}{variable name: symbol or literal character string.

Works with symbols which point to strings represeting packages}
}
\description{
Overwrite :: Operator
}
\examples{
## shorthand namespacing
d <- "dplyr"
d::select

## regular namespacing still works
dplyr::select
"dplyr"::select
}
