% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R, R/importas.R
\name{\%as\%}
\alias{\%as\%}
\alias{as}
\alias{importas}
\title{Define alias(es) to package(s)}
\usage{
pkg \%as\% nm

importas(..., .envir = parent.frame())
}
\arguments{
\item{pkg}{A name of package}

\item{nm}{An abbreviated name of the package}

\item{...}{Named arguments whose name is an alias to the value, i.e. package.}

\item{.envir}{An environment where to assign aliases.}
}
\description{
\verb{\%as\%} infix operator defines an alias to a package.
In order to define aliases for multiple packages at once, use \code{importas()}.
}
\examples{
  graphics \%as\% gr
  gr$plot

  importas(gr = graphics, st = stats)
  gr$plot
  st$median

}
