% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importAs}
\alias{importAs}
\title{Represent A Package With A Symbol}
\usage{
importAs(pkg = NULL, as = NULL)
}
\arguments{
\item{pkg}{(character) name of a (hopefully installed) package}

\item{as}{(character) symbol into which the package name should be assigned}
}
\value{
the symbol represented by \code{as} is assigned the character string \code{pkg}
}
\description{
This is merely a convenience function which signals to a reader that shorthand
(python-esque) namespacing is occurring. This is equivalent to

\code{d <- "dplyr"}

but demonstrates that this is to be explicitly used in a shorthand namespace
prefixing operation, such as \code{d::select()}
}
\examples{
importAs("dplyr", "d")
}
